rename model analysis to rai insights
