+++ responsibleai/responsibleai/__init__.py
@@ -3,10 +3,11 @@
 
 """Responsible AI SDK package."""
 
-from responsibleai.modelanalysis import ModelAnalysis, ModelTask
+from responsibleai.modelanalysis import ModelAnalysis
+from responsibleai.rai_insights import ModelTask, RAIInsights
 
 from .__version__ import version
 
 __version__ = version
 
-__all__ = ["ModelAnalysis", "ModelTask"]
+__all__ = ['ModelAnalysis', 'ModelTask', 'RAIInsights']
