+++ responsibleai/responsibleai/_managers/explainer_manager.py
@@ -311,13 +311,13 @@ class ExplainerManager(BaseManager):
             json.dump(meta, file)
 
     @staticmethod
-    def _load(path, model_analysis):
+    def _load(path, rai_insights):
         """Load the ExplainerManager from the given path.
 
         :param path: The directory path to load the ExplainerManager from.
         :type path: str
-        :param model_analysis: The loaded parent ModelAnalysis.
-        :type model_analysis: ModelAnalysis
+        :param rai_insights: The loaded parent RAIInsights.
+        :type rai_insights: RAIInsights
         :return: The ExplainerManager manager after loading.
         :rtype: ExplainerManager
         """
@@ -329,18 +329,18 @@ class ExplainerManager(BaseManager):
         if explanation_path.exists():
             explanation = load_explanation(explanation_path)
             inst.__dict__[EXPLANATION] = explanation
-        inst.__dict__['_' + MODEL] = model_analysis.model
+        inst.__dict__['_' + MODEL] = rai_insights.model
 
         with open(top_dir / META_JSON, 'r') as meta_file:
             meta = meta_file.read()
         meta = json.loads(meta)
         inst.__dict__['_' + IS_RUN] = meta[IS_RUN]
-        inst.__dict__['_' + CLASSES] = model_analysis._classes
+        inst.__dict__['_' + CLASSES] = rai_insights._classes
         inst.__dict__['_' + CATEGORICAL_FEATURES] = \
-            model_analysis.categorical_features
-        target_column = model_analysis.target_column
-        train = model_analysis.train.drop(columns=[target_column])
-        test = model_analysis.test.drop(columns=[target_column])
+            rai_insights.categorical_features
+        target_column = rai_insights.target_column
+        train = rai_insights.train.drop(columns=[target_column])
+        test = rai_insights.test.drop(columns=[target_column])
         inst.__dict__[U_INITIALIZATION_EXAMPLES] = train
         inst.__dict__[U_EVALUATION_EXAMPLES] = test
         inst.__dict__['_' + FEATURES] = list(train.columns)
