+++ responsibleai/responsibleai/modelanalysis/constants/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) Microsoft Corporation
+# Licensed under the MIT License.
+
+"""Defines ModelTask for backcompat by reimporting from new location."""
+
+from responsibleai.rai_insights.constants import ModelTask
+
+__all__ = ["ModelTask"]
