+++ responsibleai/responsibleai/rai_insights/__init__.py
@@ -1,0 +1,9 @@
+# Copyright (c) Microsoft Corporation
+# Licensed under the MIT License.
+
+"""Implementation of Model Analysis API."""
+
+from responsibleai.rai_insights.constants import ModelTask
+from responsibleai.rai_insights.rai_insights import RAIInsights
+
+__all__ = ['ModelTask', 'RAIInsights']
