+++ responsibleai/tests/counterfactual/test_counterfactual_advanced_features.py
@@ -4,7 +4,7 @@
 
 import pytest
 
-from responsibleai import ModelAnalysis
+from responsibleai import RAIInsights
 
 from ..common_utils import create_iris_data, create_lightgbm_classifier
 
@@ -22,7 +22,7 @@ class TestCounterfactualAdvancedFeatures(object):
         X_train['target'] = y_train
         X_test['target'] = y_test
 
-        model_analysis = ModelAnalysis(
+        rai_insights = RAIInsights(
             model=model,
             train=X_train,
             test=X_test.iloc[0:10],
@@ -34,13 +34,13 @@ class TestCounterfactualAdvancedFeatures(object):
         else:
             features_to_vary = [feature_names[0]]
 
-        model_analysis.counterfactual.add(
+        rai_insights.counterfactual.add(
             total_CFs=10, desired_class=2,
             features_to_vary=features_to_vary,
             feature_importance=feature_importance)
-        model_analysis.counterfactual.compute()
+        rai_insights.counterfactual.compute()
 
-        cf_obj = model_analysis.counterfactual.get()[0]
+        cf_obj = rai_insights.counterfactual.get()[0]
         assert cf_obj is not None
 
     @pytest.mark.parametrize('feature_importance', [True, False])
@@ -52,21 +52,21 @@ class TestCounterfactualAdvancedFeatures(object):
         X_train['target'] = y_train
         X_test['target'] = y_test
 
-        model_analysis = ModelAnalysis(
+        rai_insights = RAIInsights(
             model=model,
             train=X_train,
             test=X_test.iloc[0:10],
             target_column='target',
             task_type='classification')
 
-        model_analysis.counterfactual.add(
+        rai_insights.counterfactual.add(
             total_CFs=10, desired_class=2,
             features_to_vary=[feature_names[0]],
             permitted_range={feature_names[0]: [2.0, 5.0]},
             feature_importance=feature_importance)
-        model_analysis.counterfactual.compute()
+        rai_insights.counterfactual.compute()
 
-        cf_obj = model_analysis.counterfactual.get()[0]
+        cf_obj = rai_insights.counterfactual.get()[0]
         assert cf_obj is not None
 
     def test_counterfactual_manager_save_load(self, tmpdir):
@@ -77,31 +77,31 @@ class TestCounterfactualAdvancedFeatures(object):
         X_train['target'] = y_train
         X_test['target'] = y_test
 
-        model_analysis = ModelAnalysis(
+        rai_insights = RAIInsights(
             model=model,
             train=X_train,
             test=X_test.iloc[0:10],
             target_column='target',
             task_type='classification')
 
-        model_analysis.counterfactual.add(
+        rai_insights.counterfactual.add(
             total_CFs=10, desired_class=2,
             features_to_vary=[feature_names[0]],
             permitted_range={feature_names[0]: [2.0, 5.0]})
-        model_analysis.counterfactual.add(
+        rai_insights.counterfactual.add(
             total_CFs=10, desired_class=1,
             features_to_vary=[feature_names[0]],
             permitted_range={feature_names[0]: [2.0, 5.0]})
-        model_analysis.counterfactual.compute()
+        rai_insights.counterfactual.compute()
 
-        assert len(model_analysis.counterfactual.get()) == 2
-        cf_obj = model_analysis.counterfactual.get()[0]
+        assert len(rai_insights.counterfactual.get()) == 2
+        cf_obj = rai_insights.counterfactual.get()[0]
         assert cf_obj is not None
 
         save_dir = tmpdir.mkdir('save-dir')
-        model_analysis.save(save_dir)
-        model_analysis_copy = ModelAnalysis.load(save_dir)
+        rai_insights.save(save_dir)
+        rai_insights_copy = RAIInsights.load(save_dir)
 
-        assert len(model_analysis_copy.counterfactual.get()) == 2
-        cf_obj = model_analysis_copy.counterfactual.get()[0]
+        assert len(rai_insights_copy.counterfactual.get()) == 2
+        cf_obj = rai_insights_copy.counterfactual.get()[0]
         assert cf_obj is not None
