+++ responsibleai/tests/explainer_manager_validator.py
@@ -10,26 +10,26 @@ from responsibleai.exceptions import UserConfigValidationException
 LIGHTGBM_METHOD = 'mimic.lightgbm'
 
 
-def setup_explainer(model_analysis, add_explainer=True):
+def setup_explainer(rai_insights, add_explainer=True):
     if add_explainer:
-        if model_analysis.model is None:
+        if rai_insights.model is None:
             with pytest.raises(
                     UserConfigValidationException,
                     match='Model is required for model explanations'):
-                model_analysis.explainer.add()
+                rai_insights.explainer.add()
             return
         else:
-            model_analysis.explainer.add()
+            rai_insights.explainer.add()
         # Validate calling add multiple times prints a warning
         with pytest.warns(UserWarning):
-            model_analysis.explainer.add()
-    model_analysis.explainer.compute()
+            rai_insights.explainer.add()
+    rai_insights.explainer.compute()
 
 
-def validate_explainer(model_analysis, X_train, X_test, classes):
-    if model_analysis.model is None:
+def validate_explainer(rai_insights, X_train, X_test, classes):
+    if rai_insights.model is None:
         return
-    explanations = model_analysis.explainer.get()
+    explanations = rai_insights.explainer.get()
     assert isinstance(explanations, list)
     assert len(explanations) == 1
     explanation = explanations[0]
@@ -42,7 +42,7 @@ def validate_explainer(model_analysis, X_train, X_test, classes):
         assert len(explanation.local_importance_values) == len(X_test)
         assert len(explanation.local_importance_values[0]) == num_cols
 
-    properties = model_analysis.explainer.list()
+    properties = rai_insights.explainer.list()
     assert properties[ListProperties.MANAGER_TYPE] == ManagerNames.EXPLAINER
     assert 'id' in properties
     assert properties['method'] == LIGHTGBM_METHOD
@@ -55,5 +55,5 @@ def validate_explainer(model_analysis, X_train, X_test, classes):
     assert properties['is_engineered'] is False
 
     # Check the internal state of explainer manager
-    assert model_analysis.explainer._is_added
-    assert model_analysis.explainer._is_run
+    assert rai_insights.explainer._is_added
+    assert rai_insights.explainer._is_run
